AUT-1567: Implement lambda invoker in the audience loader
If there are items returned by the dynamo db query, then we need to re-invoke the lambda because there might be items still to process. We reinvoke with a last evaluated key of the final item we processed. If there are no items returned by the lambda, we know we can finish and so do not need to reinvoke.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>